refactor: move envelope helpers from patches.py to errors.py
